Refactoring consensus and stake manager
