+++ libs/ledger/include/ledger/consensus/stake_manager_interface.hpp
@@ -36,10 +36,7 @@ public:
 
   /// @name Stake Manager Interface
   /// @{
-  virtual void     UpdateCurrentBlock(Block const &current)                                = 0;
-  virtual uint64_t GetBlockGenerationWeight(Block const &previous, Address const &address) = 0;
-  virtual bool     ShouldGenerateBlock(Block const &previous, Address const &address)      = 0;
-  virtual bool     ValidMinerForBlock(Block const &previous, Address const &address)       = 0;
+  virtual void UpdateCurrentBlock(Block const &current) = 0;
   /// @}
 
 private:
