+++ libs/ledger/src/consensus/consensus.cpp
@@ -16,6 +16,9 @@
 //
 //------------------------------------------------------------------------------
 
+#include "core/random/lcg.hpp"
+
+#include <random>
 #include <utility>
 
 #include "ledger/consensus/consensus.hpp"
@@ -27,8 +30,32 @@ using NextBlockPtr    = Consensus::NextBlockPtr;
 using Status          = Consensus::Status;
 using StakeManagerPtr = Consensus::StakeManagerPtr;
 
+namespace {
+using DRNG = fetch::random::LinearCongruentialGenerator;
+
+std::size_t SafeDecrement(std::size_t value, std::size_t decrement)
+{
+  if (decrement >= value)
+  {
+    return 0;
+  }
+  else
+  {
+    return value - decrement;
+  }
+}
+
+template <typename T>
+void DeterministicShuffle(T &container, uint64_t entropy)
+{
+  DRNG rng(entropy);
+  std::shuffle(container.begin(), container.end(), rng);
+}
+}  // namespace
+
 Consensus::Consensus(StakeManagerPtr stake, BeaconServicePtr beacon, MainChain const &chain,
-                     Identity mining_identity, uint64_t aeon_period, uint64_t max_committee_size)
+                     Identity mining_identity, uint64_t aeon_period, uint64_t max_committee_size,
+                     uint32_t block_interval_ms)
   : stake_{std::move(stake)}
   , beacon_{std::move(beacon)}
   , chain_{chain}
@@ -36,9 +63,138 @@ Consensus::Consensus(StakeManagerPtr stake, BeaconServicePtr beacon, MainChain c
   , mining_address_{Address(mining_identity_)}
   , aeon_period_{aeon_period}
   , max_committee_size_{max_committee_size}
+  , block_interval_ms_{block_interval_ms}
 {
   assert(stake_);
   FETCH_UNUSED(chain_);
+}
+
+Consensus::CommitteePtr Consensus::GetCommittee(Block const &previous)
+{
+  // Calculate the last relevant snapshot
+  uint64_t const last_snapshot =
+      previous.body.block_number - (previous.body.block_number % aeon_period_);
+
+  // Invalid to request a committee too far ahead in time
+  assert(committee_history_.find(last_snapshot) != committee_history_.end());
+
+  if (committee_history_.find(last_snapshot) == committee_history_.end())
+  {
+    FETCH_LOG_INFO(LOGGING_NAME,
+                   "No committee history found for block: ", previous.body.block_number, " AKA ",
+                   last_snapshot);
+    // TODO(jmw): Check: should return nullptr here?
+  }
+
+  // If the last committee was the valid committee, return this. Otherwise, deterministically
+  // shuffle the committee using the random beacon
+  if (last_snapshot == previous.body.block_number)
+  {
+    return committee_history_.at(last_snapshot);
+  }
+  else
+  {
+    CommitteePtr committee_ptr = committee_history_[last_snapshot];
+    assert(!committee_ptr->empty());
+
+    Committee committee_copy = *committee_ptr;
+
+    DeterministicShuffle(committee_copy, previous.body.entropy);
+
+    return std::make_shared<Committee>(committee_copy);
+  }
+}
+
+bool Consensus::ValidMinerForBlock(Block const &previous, Address const &address)
+{
+  auto const committee = GetCommittee(previous);
+
+  if (!committee || committee->empty())
+  {
+    FETCH_LOG_WARN(LOGGING_NAME, "Unable to determine committee for block validation");
+    return false;
+  }
+
+  return std::find_if((*committee).begin(), (*committee).end(),
+                      [&address](Identity const &identity) {
+                        return address == Address(identity);
+                      }) != (*committee).end();
+}
+
+uint64_t Consensus::GetBlockGenerationWeight(Block const &previous, Address const &address)
+{
+  auto const committee = GetCommittee(previous);
+
+  if (!committee)
+  {
+    FETCH_LOG_WARN(LOGGING_NAME, "Unable to determine block generation weight");
+    return 0;
+  }
+
+  std::size_t weight{committee->size()};
+
+  // TODO(EJF): Depending on the committee sizes this would need to be improved
+  for (auto const &member : *committee)
+  {
+    if (address == Address(member))
+    {
+      break;
+    }
+
+    weight = SafeDecrement(weight, 1);
+  }
+
+  // Note: weight must always be non zero (indicates failure/not in committee)
+  return weight;
+}
+
+bool Consensus::ShouldGenerateBlock(Block const &previous, Address const &address)
+{
+  FETCH_LOG_DEBUG(LOGGING_NAME, "Should generate block? Prev: ", previous.body.block_number);
+
+  auto const committee = GetCommittee(previous);
+
+  if (!committee || committee->empty())
+  {
+    FETCH_LOG_WARN(LOGGING_NAME, "Unable to determine committee for block generation");
+    return false;
+  }
+
+  // At this point the miner will decide if they should produce a block. The first miner in the
+  // committee will wait until block_interval after the block at the HEAD of the chain, the second
+  // miner 2*block_interval and so on.
+  uint32_t time_to_wait = block_interval_ms_;
+  bool     in_committee = false;
+
+  bool found = false;
+
+  for (std::size_t i = 0; i < (*committee).size(); ++i)
+  {
+    FETCH_LOG_DEBUG(LOGGING_NAME, "Block: ", previous.body.block_number,
+                    " Saw committee member: ", Address((*committee)[i]).address().ToBase64(),
+                    "we are: ", address.address().ToBase64());
+
+    if (Address((*committee)[i]) == address)
+    {
+      in_committee = true;
+      found        = true;
+    }
+
+    if (!found)
+    {
+      time_to_wait += block_interval_ms_;
+    }
+  }
+
+  uint64_t time_now_ms           = static_cast<uint64_t>(std::time(nullptr)) * 1000;
+  uint64_t desired_time_for_next = (previous.first_seen_timestamp * 1000) + time_to_wait;
+
+  if (in_committee && desired_time_for_next <= time_now_ms)
+  {
+    return true;
+  }
+
+  return false;
 }
 
 void Consensus::UpdateCurrentBlock(Block const &current)
@@ -73,9 +229,11 @@ void Consensus::UpdateCurrentBlock(Block const &current)
     last_committee_created_ = int64_t(current_block_number_);
 
     CabinetMemberList cabinet_member_list;
-    auto const        current_stakers = stake_->GetCommittee(current_block_);
+    committee_history_[current.body.block_number] = stake_->BuildCommittee(current_block_);
 
-    for (auto const &staker : *current_stakers)
+    TrimToSize(committee_history_, HISTORY_LENGTH);
+
+    for (auto const &staker : *committee_history_[current.body.block_number])
     {
       FETCH_LOG_DEBUG(LOGGING_NAME, "Adding staker: ", staker.identifier().ToBase64());
       cabinet_member_list.insert(staker);
@@ -110,7 +268,7 @@ NextBlockPtr Consensus::GenerateNextBlock()
   }
 
   // Note here the previous block's entropy determines miner selection
-  if (!stake_->ShouldGenerateBlock(current_block_, mining_address_))
+  if (!ShouldGenerateBlock(current_block_, mining_address_))
   {
     return ret;
   }
@@ -121,7 +279,7 @@ NextBlockPtr Consensus::GenerateNextBlock()
   ret->body.block_number  = block_number;
   ret->body.miner         = mining_address_;
   ret->body.entropy       = entropy;
-  ret->weight             = stake_->GetBlockGenerationWeight(current_block_, mining_address_);
+  ret->weight             = GetBlockGenerationWeight(current_block_, mining_address_);
 
   return ret;
 }
@@ -168,13 +326,13 @@ Status Consensus::ValidBlock(Block const &previous, Block const &current)
   }
 
   // Here we assume the entropy is correct and proceed to verify against the stake subsystem
-  if (!stake_->ValidMinerForBlock(previous, current.body.miner))
+  if (!ValidMinerForBlock(previous, current.body.miner))
   {
     FETCH_LOG_WARN(LOGGING_NAME, "Received block has incorrect miner identity set");
     return Status::NO;
   }
 
-  if (current.weight != stake_->GetBlockGenerationWeight(previous, current.body.miner) &&
+  if (current.weight != GetBlockGenerationWeight(previous, current.body.miner) &&
       current.weight != 0)
   {
     FETCH_LOG_WARN(LOGGING_NAME,
@@ -183,6 +341,16 @@ Status Consensus::ValidBlock(Block const &previous, Block const &current)
   }
 
   return Status::YES;
+}
+
+void Consensus::Reset(StakeSnapshot const &snapshot)
+{
+  committee_history_[0] = stake_->Reset(snapshot);
+
+  if (committee_history_.find(0) == committee_history_.end())
+  {
+    FETCH_LOG_INFO(LOGGING_NAME, "No committee history found for block when resetting.");
+  }
 }
 
 void Consensus::Refresh()
