+++ libs/ledger/src/consensus/stake_manager.cpp
@@ -16,7 +16,6 @@
 //
 //------------------------------------------------------------------------------
 
-#include "core/random/lcg.hpp"
 #include "ledger/chain/block.hpp"
 #include "ledger/consensus/entropy_generator_interface.hpp"
 #include "ledger/consensus/stake_manager.hpp"
@@ -25,58 +24,13 @@
 #include <algorithm>
 #include <cstddef>
 #include <iterator>
-#include <random>
 
 namespace fetch {
 namespace ledger {
-namespace {
-
 constexpr char const *LOGGING_NAME = "StakeMgr";
-using DRNG                         = random::LinearCongruentialGenerator;
 
-std::size_t SafeDecrement(std::size_t value, std::size_t decrement)
-{
-  if (decrement >= value)
-  {
-    return 0;
-  }
-  else
-  {
-    return value - decrement;
-  }
-}
-
-template <typename T>
-void TrimToSize(T &container, uint64_t max_allowed)
-{
-  if (container.size() >= max_allowed)
-  {
-    auto const num_to_remove = container.size() - max_allowed;
-
-    if (num_to_remove > 0)
-    {
-      auto end = container.begin();
-      std::advance(end, static_cast<std::ptrdiff_t>(num_to_remove));
-
-      container.erase(container.begin(), end);
-    }
-  }
-}
-
-template <typename T>
-void DeterministicShuffle(T &container, uint64_t entropy)
-{
-  DRNG rng(entropy);
-  std::shuffle(container.begin(), container.end(), rng);
-}
-
-}  // namespace
-
-StakeManager::StakeManager(uint64_t committee_size, uint32_t block_interval_ms,
-                           uint64_t snapshot_validity_periodicity)
+StakeManager::StakeManager(uint64_t committee_size)
   : committee_size_{committee_size}
-  , snapshot_validity_periodicity_{snapshot_validity_periodicity}
-  , block_interval_ms_{block_interval_ms}
 {}
 
 void StakeManager::UpdateCurrentBlock(Block const &current)
@@ -99,155 +53,38 @@ void StakeManager::UpdateCurrentBlock(Block const &current)
     current_block_index_ = current.body.block_number;
   }
 
-  // If this would create a new committee save this snapshot
-  if (current.body.block_number % snapshot_validity_periodicity_ == 0)
-  {
-    auto snapshot = LookupStakeSnapshot(current.body.block_number);
-    committee_history_[current.body.block_number] =
-        snapshot->BuildCommittee(current.body.entropy, committee_size_);
-  }
-
   TrimToSize(stake_history_, HISTORY_LENGTH);
-  TrimToSize(committee_history_, HISTORY_LENGTH);
 }
 
-bool StakeManager::ShouldGenerateBlock(Block const &previous, Address const &address)
+StakeManager::CommitteePtr StakeManager::BuildCommittee(Block const &current)
 {
-  FETCH_LOG_DEBUG(LOGGING_NAME, "Should generate block? Prev: ", previous.body.block_number);
-
-  auto const committee = GetCommittee(previous);
-
-  if (!committee || committee->empty())
-  {
-    FETCH_LOG_WARN(LOGGING_NAME, "Unable to determine committee for block generation");
-    return false;
-  }
-
-  // At this point the miner will decide if they should produce a block. The first miner in the
-  // committee will wait until block_interval after the block at the HEAD of the chain, the second
-  // miner 2*block_interval and so on.
-  uint32_t time_to_wait = block_interval_ms_;
-  bool     in_committee = false;
-
-  bool found = false;
-
-  for (std::size_t i = 0; i < (*committee).size(); ++i)
-  {
-    FETCH_LOG_DEBUG(LOGGING_NAME, "Block: ", previous.body.block_number,
-                    " Saw committee member: ", Address((*committee)[i]).address().ToBase64(),
-                    "we are: ", address.address().ToBase64());
-
-    if (Address((*committee)[i]) == address)
-    {
-      in_committee = true;
-      found        = true;
-    }
-
-    if (!found)
-    {
-      time_to_wait += block_interval_ms_;
-    }
-  }
-
-  uint64_t time_now_ms           = static_cast<uint64_t>(std::time(nullptr)) * 1000;
-  uint64_t desired_time_for_next = (previous.first_seen_timestamp * 1000) + time_to_wait;
-
-  if (in_committee && desired_time_for_next <= time_now_ms)
-  {
-    return true;
-  }
-
-  return false;
+  auto snapshot = LookupStakeSnapshot(current.body.block_number);
+  return snapshot->BuildCommittee(current.body.entropy, committee_size_);
 }
 
-StakeManager::CommitteePtr StakeManager::GetCommittee(Block const &previous)
+StakeManager::CommitteePtr StakeManager::Reset(StakeSnapshot const &snapshot)
 {
-  // Calculate the last relevant snapshot
-  uint64_t const last_snapshot =
-      previous.body.block_number - (previous.body.block_number % snapshot_validity_periodicity_);
-
-  // Invalid to request a committee too far ahead in time
-  assert(committee_history_.find(last_snapshot) != committee_history_.end());
-
-  if (committee_history_.find(last_snapshot) == committee_history_.end())
-  {
-    FETCH_LOG_INFO(LOGGING_NAME,
-                   "No committee history found for block: ", previous.body.block_number, " AKA ",
-                   last_snapshot);
-  }
-
-  // If the last committee was the valid committee, return this. Otherwise, deterministically
-  // shuffle the committee using the random beacon
-  if (last_snapshot == previous.body.block_number)
-  {
-    return committee_history_.at(last_snapshot);
-  }
-  else
-  {
-    CommitteePtr committee_ptr = committee_history_[last_snapshot];
-    assert(!committee_ptr->empty());
-
-    Committee committee_copy = *committee_ptr;
-
-    DeterministicShuffle(committee_copy, previous.body.entropy);
-
-    return std::make_shared<Committee>(committee_copy);
-  }
+  return ResetInternal(std::make_shared<StakeSnapshot>(snapshot));
 }
 
-uint64_t StakeManager::GetBlockGenerationWeight(Block const &previous, Address const &address)
+StakeManager::CommitteePtr StakeManager::Reset(StakeSnapshot &&snapshot)
 {
-  auto const committee = GetCommittee(previous);
-
-  if (!committee)
-  {
-    FETCH_LOG_WARN(LOGGING_NAME, "Unable to determine block generation weight");
-    return 0;
-  }
-
-  std::size_t weight{committee->size()};
-
-  // TODO(EJF): Depending on the committee sizes this would need to be improved
-  for (auto const &member : *committee)
-  {
-    if (address == Address(member))
-    {
-      break;
-    }
-
-    weight = SafeDecrement(weight, 1);
-  }
-
-  // Note: weight must always be non zero (indicates failure/not in committee)
-  return weight;
+  return ResetInternal(std::make_shared<StakeSnapshot>(std::move(snapshot)));
 }
 
-void StakeManager::Reset(StakeSnapshot const &snapshot)
-{
-  ResetInternal(std::make_shared<StakeSnapshot>(snapshot));
-}
-
-void StakeManager::Reset(StakeSnapshot &&snapshot)
-{
-  ResetInternal(std::make_shared<StakeSnapshot>(std::move(snapshot)));
-}
-
-void StakeManager::ResetInternal(StakeSnapshotPtr &&snapshot)
+StakeManager::CommitteePtr StakeManager::ResetInternal(StakeSnapshotPtr &&snapshot)
 {
   // history
   stake_history_.clear();
   stake_history_[0] = snapshot;
 
-  committee_history_[0] = snapshot->BuildCommittee(0, committee_size_);
-
-  if (committee_history_.find(0) == committee_history_.end())
-  {
-    FETCH_LOG_INFO(LOGGING_NAME, "No committee history found for block when resetting.");
-  }
+  CommitteePtr new_committee = snapshot->BuildCommittee(0, committee_size_);
 
   // current
   current_             = std::move(snapshot);
   current_block_index_ = 0;
+
+  return new_committee;
 }
 
 StakeManager::StakeSnapshotPtr StakeManager::LookupStakeSnapshot(BlockIndex block)
@@ -274,27 +111,6 @@ StakeManager::StakeSnapshotPtr StakeManager::LookupStakeSnapshot(BlockIndex bloc
       return (--upper_bound)->second;
     }
   }
-}
-
-bool StakeManager::ValidMinerForBlock(Block const &previous, Address const &address)
-{
-  auto const committee = GetCommittee(previous);
-
-  if (!committee || committee->empty())
-  {
-    FETCH_LOG_WARN(LOGGING_NAME, "Unable to determine committee for block validation");
-    return false;
-  }
-
-  return std::find_if((*committee).begin(), (*committee).end(),
-                      [&address](Identity const &identity) {
-                        return address == Address(identity);
-                      }) != (*committee).end();
-}
-
-uint32_t StakeManager::BlockInterval()
-{
-  return block_interval_ms_;
 }
 
 }  // namespace ledger
