+++ libs/ledger/src/genesis_file_creator.cpp
@@ -302,7 +302,7 @@ void GenesisFileCreator::LoadConsensus(Variant const &object)
       }
     }
 
-    consensus_->stake()->Reset(*snapshot);
+    consensus_->Reset(*snapshot);
   }
   else
   {
